charge_state_v2: do not overwrite manual control mode
charge control mode is default to CHARGE_CONTROL_NORMAL, and
otherwise if we have manually command it. Manual charge mode is
usually used in the factory or faft test to maintain the battery
SoC to a specific state. However, if the AC status changed,
the charger_task will overwrite the manual control state back to
CHARGE_CONTROL_NORMAL, and this is unexpected. We should wait
for the manual control to write it back.

BUG=b:206601685
TEST=chgstate discharge, replug AC and the battery is discharging
BRANCH=cherry

Change-Id: I986ded1003baa193af15da17d1ceac44ef93294f
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>