zephyr: zmake: Remove the hard-coded list of projects in firmware_builder.py
This is actually just duplicate work of what the test phase is going
to end up doing, as testall will want to do the build anyway.

For now (that is, until we improve the zmake CLI enough to distinctly
separate the phases logically), let's just do nothing here.

BUG=b:217788621,b:218868887
BRANCH=none
TEST=CQ passes

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I91cbdf45b80a4eec1ead03a8342790c476b36585