usbc: add CONFIG_USB_PD_STARTUP_DELAY_MS
This option provides a kludge to work around issues when programming some
ECs without a battery, where the EC is allowed to come out of reset and
put into a programming mode via an external command and in doing its
normal startup may interrupt power to the board before it is put into
programming mode.

This option is enabled on Nereid where we've been having this problem,
and set to two seconds because one second didn't help with programming
failures but two seconds seems to be long enough.

BUG=b:218618254
TEST=./util/flash_ec now works on Nereid when no battery is connected
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I8d47ce440d694f90ca0f29583a2e57fcab422f31