mt8186: enable AP watchdog interrupt
Reset AP (S0->S0) when AP watchdog interrupt is raised.
Note that we should only do watchdog reset when it's a real WDT.

A real WDT is that AP_WDTRST_L is the first power signal asserts,
otherwise it's a fake WDT, because it was asserts due to AP shutdown
the WDT rail while processing AP shutdown or AP reset.

BUG=b:218420108
TEST=in EC console
     apshutdown       : S0->G3
     apreset          : S0->S0
     lid open         : G3->S0
     toggle cold reset: G3->S0
     powerbtn 8.2s    : S0->G3
     in AP console
     ectool apreset   : S0->S0
     shutdown         : S0->G3
     stop daisydog; echo > /dev/watchdog:  S0->S0
     reboot           : S0->S0
BRANCH=none

Change-Id: Ic0360c1eff1cf25d7a28974d76af41dd0c2984cd
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>