docs: Set BOARD in quick build instructions
Step 3 of the instructions in the "Building" section indicate to run a make
command that includes `make BOARD=${BOARD}`. This assumes that ${BOARD} is
set properly but the instructions to not specify to do so. The first
paragraph does link to documents that cover this but then go on to state
that you can safely ignore them.

Fix by adding a step to specify the BOARD env var, including a brief
description of where to find this value and a safe default.

BUG=b:219081299

Change-Id: I9b0a7f9b6b64ff2b1bb5844c5fa750514ecf6d1b