acpi: Add acpi support for power participant
When there is a change in power source (AC/DC/battery),
ec sends a prochot assert signal and an sci event to AP.
On recieving them, AP must query the sci event source and
if it is power source change event, along with the assert
signal, send acpi request to read th source of power. If
power source is changed, AP must increment the sequence
and send it to EC over acpi. If the sequence number,
matches with EC's value, proc hot signal is de-asserted.

BUG=b:206057712
BRANCH=none
TEST=Build and test on brya

Signed-off-by: poornima tom <poornima.tom@intel.com>
Change-Id: Ifeba1c483d47c8df03ba52441dfabf2ef5794897