herobrine: Wait negotiated VBUS transition before enabling 5V rail
Previous, the PP5000 is enabled at GPIO init. In the dead battery
case, enabling PP5000 overloads the battery charger at the initial
5V 500mA supply. We want to defer it until the PD negotiated VBUS
transition, so the battery charger won't be overloaded.

A window of PD negotiation is created. It starts from the Type-C
state reaching Attached.SNK, and eneds when the PD contract is
created. The VBUS may be raised anytime in this window.

The current implementation is the worst case scenario. Every message
the PD negotiation is received at the last moment before timeout.
The extra time is added to compensate the delay internally, like the
decision of the DPM.

BRANCH=None
BUG=b:215726554
TEST=Booted the Herobrine board with no-battery. The 5V rail got
delayed.

Change-Id: I5529762de3495e4cb6702e887bf6ee09d1711f14
Signed-off-by: Wai-Hong Tam <waihong@google.com>