nissa: correct pin definition for PP1050 monitoring
This pin is used for analog monitoring of PP1050_PROC generated by the SoC,
not as a power-good input. An entry is added to the named-adc enumeration
because PP1050_PROC has not been used before and a legacy enum value is
required even though the code using this analog channel will be
Zephyr-first.

BUG=b:203446068
TEST=Build and flash, `adc` reports PP1050_PROC voltage
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I26e0d7d8e3dbe0a2643b789cc5c6891ada710062