i2c: move PASSTRHU_RESTRICTED check to check_i2c_params
This CL moves the check about allowing PASSTHRU_RESTRICED checks from
i2c_command_passthru() to check_i2c_params().

The rationale, is that we already have a function to check whether the
params are valid and/or allowed.

This is similar to what this CL did: https://crrev.com/c/1588492

BRANCH=none
BUG=b:135642493
TEST=make buildall passes

Change-Id: I5902feae8f5b626e10053f279cb2fe2cdbba2585
Signed-off-by: Ricardo Quesada <ricardoq@google.com>