New Variant Demonstration: Part 1: BUILD.py
Part 1 of the new variant demonstration. This CL shows adding a new
variant under the Corsola baseboard called "luigi." Luigi will be very
similar to the existing Corsola variant Kingler, so we start off by
replicating the Kingler configuration. In subsequent CL(s), we will
further customize the variant.

Note: Luigi is not a real product and purely for demo purposes.

BRANCH=None
BUG=None
TEST=zmake configure -b luigi

Signed-off-by: Tristan Honscheid <honscheid@google.com>

Change-Id: I01853f56aa23003ea84fd9a98daf461d7b10f898