nissa/nereid: don't configure BC1.2 interrupts in DT
The BC1.2 interrupt line is shared with the charger on port C0, and with
the charger and TCPC on port C1 (if present). The BC1.2 driver should
not be allowed to configure its own interrupt on that pin; the
usb_c*_interrupt functions will notify the BC1.2 code if needed.

BUG=b:220775658
TEST=???
BRANCH=none

Change-Id: Ic2762b86ce206744d3a4178075bac073fdabdd5b
Signed-off-by: Peter Marheine <pmarheine@chromium.org>