presubmit: Add warning on edit of non-zephyr code
Add presubmit to repo upload that will check that all C files in the cl
are actually used in the Zephyr EC. If any of the edited files are not
actually part of the Zephyr builds, print a warning message for the
user.

This is to satisfy the goal to prevent people from accidentally making
changes in a legacy EC driver when we've already migrated that code to
Zephyr.

I excluded the baseboard, board, and chip dirs, since they are clearly
not zephyr code. Excluded test and util since those aren't ec code. And
exclude the zephyr dir, since obviously that is zephyr code.

BRANCH=None
BUG=b:205759844
TEST=Edited several included and not-included files and made sure the
warning only applied to the non-included files.

Change-Id: Ida4f6702a1408004e3a98a057c6c472247feff25
Signed-off-by: Jeremy Bettis <jbettis@google.com>