presubmit: Add warning on edit of non-zephyr code
Add presubmit to repo upload that will check that all C files in the cl
are actually used in the Zephyr EC. If any of the edited files are not
actually part of the Zephyr builds, print a warning message for the
user.

This is to satisfy the goal to prevent people from accidently making
changes in a legacy EC driver when we've already migrated that code to
Zephyr.

BRANCH=None
BUG=b:205759844
TEST=Edited several included and not-included files and made sure the
warning only applied to the non-included files.

Change-Id: Ida4f6702a1408004e3a98a057c6c472247feff25
Signed-off-by: Jeremy Bettis <jbettis@google.com>