hook: Add RO dependency hook
This presubmit hook checks if gpio.inc of a board is
edited when CONFIG_SYSTEM_UNLOCK is undefined - an
indicator that the board's RO FW was qualified.

If so, the commit message should mention testing that board
against RO. If it doesn't, the CL won't pass this check.

BUG=b:194095346
TEST=local changes to gpio.inc, muti-commit CLs
BRANCH=None

Signed-off-by: Nehemiah Dureus <ndureus@chromium.org>
Change-Id: I49acc5eb1c1eed7fa8497bf2146dae2eafa13df5