u2f: do not commit state changes on TPM command context.
g2f_attestation_cert() is another function which is invoked on the TPM
command context, when virtual TPM NVMEM spaces are read.

One of the side effects of invoking of g2f_attestation_cert() is the
creation of the U2F state, if it did not exist before. In this case
the state should not be immediately committed to the NVMEM, the commit
will happen when the TPM command execution is completed.

BUG=b:199981251
TEST=running ./test/tpm_test/tpmtest.py does not trigger the 'attempt
     to commit in unlocked state' message any more.
     'make buildall' and 'make CRYTPO_TEST=1 BOARD=cr50' pass

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: I708e8807ffd3207cc6ab84a0e380908e715f7a15