zmake: Don't exclude anything from coverage
In zmake, don't remove any directories from the coverage reports.

In gitlab, add the exclusions back in so that we get the same finished
html report that we got before.

BRANCH=None
BUG=b:156895937
TEST=Ran firmware_builder.py, verified zephyr/test & zephyr/emul
Ran .gitlab-ci.yml commands for herobrine and verified that
build/herobrine/output/filtered_no_zephyr.info is unchanged.

Signed-off-by: Jeremy Bettis <jbettis@google.com>
Change-Id: Ie8fd1e787d8bb6e745f89d773113cccb6c335009