zephr: Add extra access for power signals
Add access APIs for 'named-gpios' and 'gpio-interrupts' to
the power signal access API.

This allows the power signal DTS config to reference the
cros-ec 'named-gpios' nodes and the 'gpio-interrupts' node.

This avoids having to duplicate the GPIO config (and to mark
named-gpios as 'no-auto-init').

The other advantage is that the named-gpios GPIOs are handled specially
during a RW sysjump (output values are maintained).

BUG=b:201000950
TEST=zmake configure -b brya
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: I49ba32ce122ecee3a2a95b56cbed78108e2d9c75