nissa/nereid: enable eSPI peripheral
Support for CONFIG_ESPI_PERIPHERAL_CUSTOM_OPCODE was enabled in the it8xxx2
eSPI driver (https://github.com/zephyrproject-rtos/zephyr/pull/41429),
so we can now enable it on Nereid.

Many of the related options are default-enabled for NPCX9 (Nivviks) but not
for it8xxx2, so we explicitly enable the relevant options for the entire
project.

BUG=b:211772002
TEST=it's broken: "Get lpc_memmap failed" loop and watchdog reset
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: Ia6f397d8a0b7d11f617b20e9983b7c1b2d62f9b6