gpio: change A1 pinmux flag to DIO_INPUT instead of GPIO_INPUT
The DIOA1 PINMUX definition uses a GPIO flag instead of a DIO flag. It
doesn't matter that much, because GPIO_INPUT maps to DIO_DIRECT_INPUT
which is a noop. i2cp_set_pinmux configures the DIOA1 input in existing
images. This change just modifies the flag for correctness.

BUG=b:221090807
TEST=check pinmux output on spi and i2c boards. Verify it doesn't
change.

Change-Id: I227156e5799d872da32a87a7bcab4ae638c18c08
Signed-off-by: Mary Ruthven <mruthven@chromium.org>