ccd_config: allow the option to restrict ccd in prepvt images
By default ccd open can be sent from the console in normal mode with
prepvt images. The open capabilities are set to Always which should
determine that open is allowed, but prepvt images completely ignore the
ccd settings. This change modifies the CCD_OPEN_PREPVT behavior to
honor the capabilities, so someone could restrict ccd open in prepvt
images with the ccd capability settings.

BUG=b:221260041
TEST=manual see bug

Change-Id: I1c3fc4f5be27a08ea9071966cc01c4b9ff20dbe5
Signed-off-by: Mary Ruthven <mruthven@chromium.org>