zephyr: test enabling charge state debugging
Verify that the charging progress was printed. This is not 100% ideal,
since it relies on an additional variable being set in order to expose
whether or not the function has been called. Ideally, we'd be watching
the logs in the test and verifying that the correct information was
printed, but we do not yet have that capability.

BRANCH=none
BUG=b:221471404
TEST=zmake configure --clobber --test test-drivers

Signed-off-by: Yuval Peress <peress@google.com>
Change-Id: Iec1dd156ce7711f7550b142c528bd2374ed99258