zephyr: shim: move the PWM HC implementation in the shim driver
We are moving in the direction of removing the named-pwm implementation,
and using the standard Zephyr shell command for troubleshooting. This
leaves the PWM host command, which is used for keyboard and display
backlight control. The implementation was getty fairly cluttered in the
common code, and would get even worse with the generic channel
compatibility code.

Move an Zephyr only implementation in the shim, with just keyboard
backlight support for now, so that we have a simpler code base to work
with.

BRANCH=none
BUG=b:217741090
TEST=zmake testall
TEST=ectool pwmsetduty on volteer

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Change-Id: Ic6295c0ac4b1b572d81851a8f518ca13a9a4daf5