zephyr: Decouple task start in AP power sequence
Decouple the task setup and start from the main power sequence
code so that the task can be started as an EC task, or as a zephyr task.
The big difference is that the EC task is started as part of the
rest of the EC task initialisation so that initialisation
dependencies can be correctly met.

BUG=b:222937759
TEST=zmake configure -b nereid; flash and run
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: Iacc438283b8639f02cab661b08950dfc9212aa6b