nissa: make EC_SOC_INT_ODL default-idle
Its default state was set to low (asserted), probably because the
difference between OUTPUT_ODL and OUTPUT_ODR in the pinmap isn't very
obvious (this error has already been fixed for some other pins). Correct
it to be default-idle so the AP doesn't need to waste a lot of time
servicing interrupts about nothing.

BUG=b:222023305
TEST=builds
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I07ee9f0b2ef90d11a4f9679e246cb1d840defa28