npcx: adc: support faster ADC conversion speed
In the ADC keyboard scan application, it requires higher ADC sampling
time to meet the keyboard scan timing specification.
To meet the timing constraint, this CL increases the ADC clock frequency
from ~2 to 7.5 MHz.
This change can apply to npcx7 and later chips because the basic clock
(APB1) of the ADC module is at 15 MHz.
For npcx5, the ADC clock frequency keeps at ~2 MHz because the APB1
clock is at (15/4) MHz.

BUG=b:208773873
BRANCH=none
test=read 8 ADC channels in board npcx9_evb via console command "adc";
it spent ~0.92 milli-seconds to read 8 ADC channels;
test = "make buildall"

Signed-off-by: Jun Lin <CHLin56@nuvoton.com>
Change-Id: Ife59f667b1aafc926b2ad05209e7cc6185b82f21