battery/smart: Don't update fields if reading fails
Currently, battery_get_params clears fields which failed to be read
because it starts with batt_new = {0} and ends with memcpy(batt,
batt_new. This causes a remaining capacity to be reported as 0 when
there is a glitch in the communication with a battery.

This patch fixes the issue by copying the current value to batt_new.

BUG=b:217401040,b:222722611
BRANCH=None
TEST=Guybrush

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: I5b69b708d149d1c750551d8cf957eab3fdc8aa64