battery/smart: Don't update fields if reading fails
Currently, battery_get_params starts with batt_new = {0} and ends with
memcpy(batt, batt_new). That is, if battery_get_params fails to read a
field, we set a BATT_FLAG_BAD_* and clear the field. This behavior
forces us to check the flags everywhere fields are read (to
distinguish 0 v.s. invalid 0). It also makes EC forget the last known
good values.

Instead, we will always keep good values in the global batt_params.
That is, we update fields only if reads are successful. (That's how
other implementations of battery_get_params behave.) Thus, if
display_soc is zero, it was computed legitimately as zero (except the
very initial value).

BUG=b:217401040,b:222722611
BRANCH=None
TEST=Guybrush

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: I5b69b708d149d1c750551d8cf957eab3fdc8aa64