zephyr: emul: Add support for RECEIVE_DETECT in TCPCI
TCPCI handle RECEIVE_DETECT register properly. TCPM should set this
register only when it is ready to receive this certain kind of messages.
Partner emulators are notified when send fail because of RECEIVE_DETECT
value. In real world it would be the same as not receiving GoodCRC
message from partner point of view.

BUG=b:224927648
BRANCH=none
TEST=zmake configure --test test-drivers

Signed-off-by: Tomasz Michalec <tm@semihalf.com>
Change-Id: Ia6f2c16e94c600a81b9071d11ea79514f02a2081