zephyr: support smart USB-A port control
This enables support for "smart" USB-A ports in addition
to the existing "dumb" support, allowing one of them to
be selected as well as implementing kconfig support for
most of the other options for "smart" ports.

For existing boards that enabled USB-A support with the default
"dumb" support, their configuration is changed to specifically
select that mode because the new choice option cannot have a
default.

BUG=b:223937974
TEST="smart" port control can be turned on for Nereid
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: If5b0fb393451d31a937c73be2aa3b9623c69307f