ap_hang_detect: Refactor ap_hang_detect for multiple timers
ap_hang_detect is currently only supports one timer running at a time.
This limits it's usefulness since some hang patterns are incompatible
with each other.

This change scales the utility to support multiple timers, each with
different hang detection criteria. This approach also allows some
portions of the implementation to be greatly simplified.

There's no current users of this utility, so it's safe to refactor it.

BUG=b:218892808
BRANCH=guybrush
TEST=Detect hang in multiple scenarios

Change-Id: Ia96f82b19145b82e355814337c8b211f46325d73
Signed-off-by: Rob Barnes <robbarnes@google.com>