zephyr: Hide gpio_enable_interrupt if no dt nodes
Move cros_ec_gpio_interrupts into the
#if DT_HAS_COMPAT_STATUS_OKAY(cros_ec_gpio_interrupts)
block, as it depends on gpio_enable_dt_interrupt, which is there.

Tests that don't define any cros_ec,gpio_interrupts fail to build on
gitlab, but succeed in the chroot for an unknown reason. This
change should make the breakages happen earlier and be more obvious.

BRANCH=None
BUG=b:222934152
TEST=zmake configure --test test-ap_power

Signed-off-by: Jeremy Bettis <jbettis@google.com>
Change-Id: I34a502be8314422113d31c0190679ff9caf03d0e