nissa/nereid: shrink hostcmd console buffer
We need more memory for the USB mux task, and this buffer is a convenient
place to get it from, at the cost of getting less history from
logs gathered by the host processor.

This change also removes the board-specific serial buffer size option,
since it's also set in prj.conf.

BUG=b:223044986,b:223749568
TEST=nereid build succeeds with USB mux task enabled
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I68db610756fd80135529cd7eda582ad1f697d042