nissa: run sub-board detection at PRIO_INIT_I2C
HOOK_PRIO_INIT_I2C is documented as the priority at which I2C is
guaranteed to be ready, and +1 from there is also the priority at
which chargers are initialized so we must use PRIO_INIT_I2C because
otherwise chargers (and maybe other USB-C peripherals) won't get
initialized.

I noticed this when Nereid failed to charge, and the "PD port count
not initialized" warning was being output. Since this is probably a
serious problem, I've also added an assertion to that function to fail
fast in debug builds.

BUG=none
TEST=nereid charges again, and does not warn about uninitialized port
     counts.
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: If6cc0de5279bc5616813a38d8b29e3b747514049