nissa: use standard keyboard column 2 invert option
On Nivviks this was being accomplished by driving the column 2 output
with an active-low GPIO, but doing this portably requires setting the
dedicated KEYBOARD_COL2_INVERTED option because Nereid's EC does not
need to drive the signal as a GPIO in order to invert it and
consequently cannot invert it in the device tree in the same way.

It appears that because this option was not enabled on Nivviks, KS02
probably wasn't actually being inverted because cros_kb_raw_set_col2()
does nothing unless both the GPIO alias is set and this option is
enabled.

BUG=b:225082447
TEST=Keyboard column 2 output now idles high on Nereid
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: Ia3e5966d642b4ffb458cb1d91015869763ba00b4