zephyr/hooks: don't store hooks in RAM
This saves 8 bytes times the number of hook handlers in RAM, which amounts
to nearly 1KB in most configurations. It trades that RAM for higher
cost in calling handlers since the hook list must be traversed multiple
times for each notification, but that matches the ECOS behavior which
implies the performance cost is acceptable.

BUG=b:223044986
TEST=zmake testall; 944 bytes of RAM are saved on Nereid
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I823f2a974faf69fa5195f11c645b569fb57854a9