kinox: add enable and disable ppc sink mode
When Kinox is sinking power from the type-c port then enable the PPC to sink and disabled when using the barrel jack.
And the Kinox does not support the adapter change when EC is on.

BUG=b:225769067
BRANCH=none
TEST=Test on Kinox,
The system can power on to OS via barrel adapter or type-c adapter.
Plugin the barrel adapter the log has the "New charger is p1" and successfully disables C0 to sink path.
Plugin the Type-C 65w adapter the log has the"New charger is p0" and successfully enables C0 to sink path.

Signed-off-by: Matt Wang <matt_wang@compal.corp-partner.google.com>
Change-Id: I7adaa113d15cd607f1be799da072386160bdcb74