fan: npcx: fine tune the fan control
The fan used on the Vell platform requires a longer response time than
others. It needs about 1 second to output the TACH signal to EC after EC
starts to toggle the PWM signal. During this period, the fan driver
adjusts the PWM duty per 200 milli-seconds and has the chance to let the
duty reach 100% with the original fan control mechanism. It makes the
user feel the fan spins unsmoothly. To bypass it, this CL steps the PWM
duty not too aggressively when two consecutive RPM reading is zero.

BUG=b:225208265
BRANCH=none
TEST=pass "make buildall"
TEST=set RPM=3000 on VELL, the fan duty doesn't reach 100% afer 1
second.

Signed-off-by: Jun Lin <CHLin56@nuvoton.com>
Change-Id: Idc14c859576247759e9119abee5fdeb1f483331d