TCPMV2: Send Alert Message on Power State Change
Some docks will have lights which are meant to reflect the current power
state of a Chromebook. This CL is to send an Alert message to partners
when power state changes. If the power state is important to the
partner, they can then request more information with a Get_Status
command.

BUG=b:226228461
TEST=Changed power state of device and observed Alert messages on a
twinkie
BRANCH=None

Signed-off-by: Jameson Thies <jthies@google.com>

Change-Id: I49720547bd881e3a0f079b135ef4177ea337d1d8