board/taeko: Override NVME and EMMC FW_CONFIG bits based on GPIO
Taeko has an EMMC_SKU_DET strap pin which signals whether or not the
current board has EMMC or NVME storage. In the factory, the CBI images
are prepared with both NVME and EMMC masks set to enabled, but this
causes an issue in the AP firmware PCIe RP initialization because the
two devices are designed to use the same CLKREQ# pin regardless of
SKU. The FSP cannot handle this, thus we are left with trying to
override these values. Thus this patch will read the FW_CONFIG value
at init time, and if either both NVME and EMMC are selected, or
neither, then it will read the EMMC_SKU_DET pin and set the two
FW_CONFIG masks accordingly.

BUG=b:224884408
BRANCH=firmware-brya-14505.B
TEST=verified by ODM

Change-Id: Ia8ad12fb785a75af6311acb41db7cedce58ad881
Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>