ec_memmap: Add Power source to EC MEMMAP region
As part of implementing Dynamic Tuning Technology
for Platform Power Participant, power source is
one of the object that requires updation from EC
side. As this needs to be read by AP, this must be
present in EC MEMMAP region. As per doc-626708, this
is an 8 bit value, comprising of sequence number
and source of power.

BUG=b:205928013
BRANCH=None
TEST=make build_cros_ec_commands

Signed-off-by: poornima tom <poornima.tom@intel.com>
Change-Id: Iebfb305e95e647cbe3d83a351765b7711070981e