nissa/nereid: configure HDMI+USB-A sub-board
This corrects some device tree settings so the GPIOs for this sub-board
can be configured correctly, hooks up AP power events to control
power to the sub-board, and hooks up HPD interrupts to pass through to
the AP.

In the device tree, some configuration needs to be moved to code because
gpio_pin_configure_dt combines the flags from device tree with those
passed to the function, with the result that configuring sb_1 as an output
as required for this sub-board ended up requesting the wrong configuration.
The I/O configurations for sb_3 and sb_4 are also corrected: sb_3 is the
HPD input (not output), and sb_4 is the VCC enable output (not an input).

BUG=b:226215993
TEST=Still kind of broken!
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I0838c4b8fab729aa09f92570c4c145271edb6a9e