power: Add PSRC support for DPTT
For DTT Power Participant, PSRC is one of the object.
PSRC is allocated in EC memory with 8 bit size. Whenever
there is a power source change, the lower nibble of this
memory must be updated with the source information and
AP must be notified about it by an sci event.
Also, for AC removal or low battery level condition,
a PROCHOT must be asserted and the upper nibble
indicating the sequence number must be incremented.
Once prochot is asserted, if there is no PBOK ACK
from AP within 2sec, PROCHOT is deasserted.
Document - #626707 has more details. This CL consists
of PSRC updation, PROCHOT assertion/deassertion in the
absence of PBOK and sci event generation.
CL also has zephyr support for the newly added config.

BUG=b:205928013
BRANCH=None
TEST=make buildall

Signed-off-by: poornima tom <poornima.tom@intel.com>
Change-Id: Id98a42a3d4c5418b89253da3da62ab91f6e83ebe