power: Add ARTG to power participant
As part of supporting power participant for
Intel Dynamic Tuning, support for adapter rating
component is added. When Power source is
DC source, this value must be zero. When USB PD
source, adapter rating must be calculated based
on the voltage and current. This is saved in
EC_MEMMAP addr - 0xa7, so that it can be read
by Bios later. As per document-#626708, artg
must be saved at bits[11:0]. Hence, at EC side,
this value is saved in 50 mW units. This needs
conversion to 1mW unit at AP level, for
sending it to DPTF.

BUG=b:205928013
BRANCH=None
TEST=make buildall

Signed-off-by: poornima tom <poornima.tom@intel.com>
Change-Id: I3e84850567000da650866a4cacafa75f9655b957