power: Add ARTG to power participant
As part of supporting power participant for
Intel Dynamic Tuning, support for adapter rating
component must be added. When Power source is
DC source, this value must be zero. When USB PD
source, adapter rating must be calculated based
on the voltage and current. This is saved in
EC_MEMMAP addr - 0xac, so that it can be read
by Bios later.

BUG=b:205928013
BRANCH=None
TEST=make buildall

Signed-off-by: poornima tom <poornima.tom@intel.com>
Change-Id: I3e84850567000da650866a4cacafa75f9655b957