TCPMV2: Respond to Get_Status messages with Status
USB-C partners may be interested in a Chromebook's power state. And,
with USB power delivery, power state information is included in the
Status message. This CL adds support to TCPMV2 to respond to Get_Status
messages with a Status message that includes the Chromebook's power
state and a corresponding power state indicator.

BUG=b:227236917
TEST=Triggered Get_Status message from partner (OWC TBT4 Dock) with
power state change alert message and observed a Status response message
on a Twinkie.
BRANCH=None

Signed-off-by: Jameson Thies <jthies@google.com>

Change-Id: I867ce6048e13ecfb9c3c99e597d4b357728636c2