COIL: update chip/max32660
This CL updates chip/max32660 according to the COIL guidelines.

It also updates some indentation issues in the comments.

But in order to have the CL as clean as possible, it does not fix all
the indentation issues that are present in modified files.
This CL was submitted with "--no-verify" to by-pass the pre-hook checks.
Needed to bypass the too-many indentation issues that are present.

BRANCH=None
BUG=b:163885307
TEST=compare_build.sh -b max32660-eval matches

Change-Id: I2aba21865b60a03ad5149359bbba6ddf91b5924c
Signed-off-by: Ricardo Quesada <ricardoq@google.com>