Felwinter: Fixed EN_PP5000_PEN doesn't deasserted when pen is plug-out
The first time we plug-out the pen, the pin EN_PP500_PEN does not
deasserted.
We need to record the initial pen detect status to avoid the
EN_PP500_PEN status doesn't change when pen status is change at the
first time

BUG=b:196009529
BRANCH=brya
TEST=ectool gpioget to check the EN_PP500_PEN status can be change when
the pen status is changed at the first time.

Signed-off-by: Josh Tsai <josh_tsai@compal.corp-partner.google.com>
Change-Id: I42bc4ecf1714919fe2193087adcb00462644b6ff